Upload files by chunks. Check md5 sum.

+++ doc/web/upload.php
@@ -16,9 +16,9 @@ function getRealIpAddr()
 	}
 	return $ip;
 }
-
-$size = $_SERVER["CONTENT_LENGTH"];
 $filename = addslashes($_SERVER["HTTP_FILENAME"]);
+$md5 = addslashes($_SERVER["HTTP_MD5"]);
+$chunknumber = addslashes($_SERVER["HTTP_CHUNK"]);
 $description = addslashes($_SERVER["HTTP_DESCR"]);
 $ip = addslashes(getRealIpAddr());
 
@@ -28,27 +28,48 @@ if (!$link) {
 }
 mysql_select_db("s210915_main");
 
-$deletequery = "DELETE FROM files WHERE filename = '". addslashes($filename) . "'";
-mysql_query($deletequery) or die("Error. Query 1 failed.\n");
+
+if ($md5) {
+	echo "Filename: $filename\n";
+	$sql = "SELECT content FROM files WHERE filename = '$filename' ORDER BY chunk";
+	$result = mysql_query($sql) or die("Can't execute query 3");
+
+	$digest = hash_init("md5");
+	$len = 0;
+	while ($row = mysql_fetch_row($result)) {
+		$len = $len + strlen($row[0]);
+		hash_update($digest, $row[0]);
+		echo "-MD5: " . md5($row[0]) . "\n";
+	}
+	if (hash_final($digest) == $md5) {
+		mysql_free_result($result);
+		$setokquery = "UPDATE files SET md5ok = 1 WHERE filename = '$filename'";
+		mysql_query($setokquery) or die("Can't execute query 4" . mysql_error());
+		echo "File $filename completed successfully. md5: $md5\n";
+	} else {
+		die("ERROR: Wrong md5 for file $filename. Expected md5: $md5 (len: $len)\n");
+	}
+	mysql_close($link);
+	exit;
+}
+
+if ($chunknumber == "1") {
+	$deletequery = "DELETE FROM files WHERE filename = '". addslashes($filename) . "'";
+	mysql_query($deletequery) or die("Error. Query 1 failed.\n");
+}
 
 $putdata = fopen("php://input", "r");
-$chunk = 1;
-while (!feof($putdata)) {
-	$content = addslashes(fread($putdata, 500000));
-	if (strlen($content) > 0) {
-		$sql = "INSERT INTO files (content, ip, description, filename, chunk)
-				VALUES ('$content', '$ip', '$description', '$filename', $chunk)";
-		$r = mysql_query($sql) or die("Error. Query 2 failed, chunk $chunk" . mysql_error() . "\n");
-		echo "Chunk $chunk: " . strlen($content) . "  ($r row added)\n";
-		$chunk++;
-	}
-}
+$content = addslashes(fread($putdata, 1000000));
+
+$sql = "INSERT INTO files (content, ip, description, filename, chunk)
+		VALUES ('$content', '$ip', '$description', '$filename', $chunknumber)";
+
+$r = mysql_query($sql) or die("Error. Query 2 failed, chunk $chunk" . mysql_error() . "\n");
+echo "Chunk $chunknumber (size=" . strlen($content) . ")\n";
 
 fclose($putdata);
 
 mysql_close($link);
-
-echo "OK. Size: $size bytes, " . ($chunk - 1) . " chunks\n";
 
 
 ?>
